@@ -20,13 +20,13 @@
     "codecov": "^1.0.1",
     "cross-spawn": "^5.1.0",
     "enzyme": "^2.8.2",
-    "eslint": "^3.17.1",
+    "eslint": "^4.1.0",
     "eslint-plugin-babel": "^4.1.1",
-    "eslint-plugin-flowtype": "^2.30.3",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-markdown": "^1.0.0-beta.4",
-    "eslint-plugin-react": "^6.7.1",
-    "flow-bin": "^0.47.0",
+    "eslint-plugin-flowtype": "^2.34.0",
+    "eslint-plugin-import": "^2.6.0",
+    "eslint-plugin-markdown": "^1.0.0-beta.6",
+    "eslint-plugin-react": "^7.1.0",
+    "flow-bin": "^0.48.0",
     "glob": "^7.1.1",
     "graceful-fs": "^4.1.11",
     "immutable": "^4.0.0-rc.1",
@@ -55,11 +55,9 @@
     "build-clean": "rm -rf ./packages/*/build ./packages/*/build-es5",
     "build": "node ./scripts/build.js",
     "clean-all": "rm -rf ./node_modules && rm -rf ./packages/*/node_modules && rm -rf ./integration_tests/*/*/node_modules && yarn run build-clean",
-    "danger": "node ./danger/node_modules/.bin/danger",
     "jest": "node ./packages/jest-cli/bin/jest.js",
     "jest-coverage": "yarn run jest --silent -- --coverage",
-    "lint": "yarn run lint-prettier --silent && eslint . --cache && yarn run lint-docs --silent",
-    "lint-docs": "eslint --config ./.eslintrc-docs.json --no-ignore --ext md ./docs/*.md",
+    "lint": "yarn run lint-prettier --silent && eslint . --cache --ext js,md",
     "lint-es5-build": "eslint --no-eslintrc --no-ignore --env=browser packages/*/build-es5",
     "lint-prettier": "node scripts/prettier.js lint",
     "prettier": "node scripts/prettier.js write",
@@ -101,7 +99,7 @@
     },
     "setupTestFrameworkScriptFile": "<rootDir>/testSetupFile.js",
     "snapshotSerializers": [
-      "<rootDir>/packages/pretty-format/build/plugins/ConvertAnsi.js"
+      "<rootDir>/packages/pretty-format/build/plugins/convert_ansi.js"
     ],
     "testEnvironment": "./packages/jest-environment-node",
     "testPathIgnorePatterns": [
@@ -113,7 +111,7 @@
       "/packages/.*/build-es5"
     ],
     "testMatch": [
-      "**/*-test.js"
+      "**/*.test.js"
     ]
   }
 }
