@@ -65,7 +65,7 @@ const cleanup = (directory: string) => rimraf.sync(directory);
  *   '/home/tmp',
  *   {
  *     'package.json': '{}',
- *     '__tests__/test-test.js': 'test("lol")',
+ *     '__tests__/test.test.js': 'test("lol")',
  *   }
  * );
  */
@@ -137,6 +137,7 @@ const cleanupStackTrace = (output: string) => {
     .replace(/\n.*at.*next_tick\.js.*$/gm, '')
     .replace(/\n.*at Promise \(<anonymous>\).*$/gm, '')
     .replace(/\n.*at <anonymous>.*$/gm, '')
+    .replace(/\n.*at Generator.next \(<anonymous>\).*$/gm, '')
     .replace(/^.*at.*[\s][\(]?(\S*\:\d*\:\d*).*$/gm, '      at $1');
 };
 
