@@ -15,14 +15,14 @@ import crypto from 'crypto';
 import fs from 'fs';
 import path from 'path';
 import jestPreset from 'babel-preset-jest';
+import {transform as babelTransform, util as babelUtil} from 'babel-core';
+import babelIstanbulPlugin from 'babel-plugin-istanbul';
 
 const BABELRC_FILENAME = '.babelrc';
 const BABELRC_JS_FILENAME = '.babelrc.js';
 const BABEL_CONFIG_KEY = 'babel';
 const PACKAGE_JSON = 'package.json';
 const THIS_FILE = fs.readFileSync(__filename);
-
-let babel;
 
 const createTransformer = (options: any) => {
   const cache = Object.create(null);
@@ -102,11 +102,7 @@ const createTransformer = (options: any) => {
       config: ProjectConfig,
       transformOptions: TransformOptions,
     ): string {
-      if (!babel) {
-        babel = require('babel-core');
-      }
-
-      if (babel.util && !babel.util.canCompile(filename)) {
+      if (babelUtil && !babelUtil.canCompile(filename)) {
         return src;
       }
 
@@ -116,7 +112,7 @@ const createTransformer = (options: any) => {
         // Copied from jest-runtime transform.js
         theseOptions.plugins = theseOptions.plugins.concat([
           [
-            require('babel-plugin-istanbul').default,
+            babelIstanbulPlugin,
             {
               // files outside `cwd` will not be instrumented
               cwd: config.rootDir,
@@ -126,7 +122,7 @@ const createTransformer = (options: any) => {
         ]);
       }
 
-      return babel.transform(src, theseOptions).code;
+      return babelTransform(src, theseOptions).code;
     },
   };
 };
