@@ -8,7 +8,10 @@
  * @flow
  */
 
+import git from './git';
+import hg from './hg';
+
 module.exports = {
-  git: require('./git'),
-  hg: require('./hg'),
+  git,
+  hg,
 };
