@@ -33,7 +33,7 @@ type Options = {|
   withAncestor?: boolean,
 |};
 
-export type PathPattern = {|
+export type TestSelectionConfig = {|
   input?: string,
   findRelatedTests?: boolean,
   lastCommit?: boolean,
@@ -216,13 +216,24 @@ class SearchSource {
     });
   }
 
-  getTestPaths(pattern: PathPattern): Promise<SearchResult> {
-    if (pattern.onlyChanged) {
-      return this.findChangedTests({lastCommit: pattern.lastCommit});
-    } else if (pattern.findRelatedTests && pattern.paths) {
-      return Promise.resolve(this.findRelatedTestsFromPattern(pattern.paths));
-    } else if (pattern.testPathPattern != null) {
-      return Promise.resolve(this.findMatchingTests(pattern.testPathPattern));
+  getTestPaths(
+    testSelectionConfig: TestSelectionConfig,
+  ): Promise<SearchResult> {
+    if (testSelectionConfig.onlyChanged) {
+      return this.findChangedTests({
+        lastCommit: testSelectionConfig.lastCommit,
+      });
+    } else if (
+      testSelectionConfig.findRelatedTests &&
+      testSelectionConfig.paths
+    ) {
+      return Promise.resolve(
+        this.findRelatedTestsFromPattern(testSelectionConfig.paths),
+      );
+    } else if (testSelectionConfig.testPathPattern != null) {
+      return Promise.resolve(
+        this.findMatchingTests(testSelectionConfig.testPathPattern),
+      );
     } else {
       return Promise.resolve({tests: []});
     }
